fix: incorrect validation in #player_input
Before fixing, user_input between '1' and '7' will be treated as valid
input. It leads some invalid inputs such as '10' will also be treated as
valid.

In this fix, the validation is updated to check if user_input contains
only number and it is between 1 and 7 after converting from string to
integer.